Restructured folder directory So I started organizing things a little bit better. See if this works for you and if so. Accept
