+++ srcs/CMakeLists.txt
@@ -2,19 +2,19 @@ project(webserv_lib)
 
 add_library(${PROJECT_NAME} STATIC
 	webserv.cpp
-	HTTPServer.cpp
-	HTTPHandler.cpp
-	HTTPParser.cpp
-	HTTPRequest.cpp
-	Config.cpp
-	ServerConfig.cpp
-	Logger.cpp
-	Error.cpp
-	Socket.cpp
-	SocketAddress.cpp
-	SocketConnection.cpp
-	TCPSocket.cpp
-	ServerSocket.cpp
+	http/HTTPServer.cpp
+	http/HTTPHandler.cpp
+	http/HTTPParser.cpp
+	http/HTTPRequest.cpp
+	server/Config.cpp
+	server/ServerConfig.cpp
+	utils/Logger.cpp
+	utils/Error.cpp
+	socket/Socket.cpp
+	socket/SocketAddress.cpp
+	socket/SocketConnection.cpp
+	socket/TCPSocket.cpp
+	socket/ServerSocket.cpp
 )
 
 add_library(42::${PROJECT_NAME} ALIAS ${PROJECT_NAME})
