+++ srcs/http/HTTPHandler.cpp
@@ -1,4 +1,4 @@
-#include "HTTPHandler.hpp"
+#include "http/HTTPHandler.hpp"
 
 namespace webserv {
