+++ srcs/utils/Logger.cpp
@@ -1,4 +1,4 @@
-#include "Logger.hpp"
+#include "utils/Logger.hpp"
 
 Logger::Logger(LogLevel l) : level(l) {}
