+++ tests/test_Error.cpp
@@ -1,4 +1,4 @@
-#include "Error.hpp"
+#include "utils/Error.hpp"
 
 #include <gtest/gtest.h>
 #include <iostream>
