+++ tests/test_HTTPParser.cpp
@@ -1,4 +1,4 @@
-#include "HTTPParser.hpp"
+#include "http/HTTPParser.hpp"
 
 #include <gtest/gtest.h>
 #include <iostream>
