+++ tests/test_HTTPServer.cpp
@@ -1,4 +1,4 @@
-#include "HTTPServer.hpp"
+#include "http/HTTPServer.hpp"
 
 #include <gtest/gtest.h>
 #include <iostream>
