+++ tests/test_Logger.cpp
@@ -1,4 +1,4 @@
-#include "Logger.hpp"
+#include "utils/Logger.hpp"
 
 #include <gtest/gtest.h>
