+++ tests/test_socket.cpp
@@ -1,6 +1,5 @@
-
 #include "Client.hpp"
-#include "TCPSocket.hpp"
+#include "socket/TCPSocket.hpp"
 
 #include <arpa/inet.h>
 #include <cerrno>
